Run visualization from Python server
