+++ backend/.gitignore
@@ -1,4 +1,6 @@
 /.direnv/
+/anim/*
+!/anim/.keep
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
