+++ backend/requirements.txt
@@ -6,13 +6,18 @@ MarkupSafe==1.1.1
 Pillow==6.0.0
 Werkzeug==0.15.4
 cycler==0.10.0
+dnspython==1.16.0
+eventlet==0.25.0
+greenlet==0.4.15
 itsdangerous==1.1.0
 kiwisolver==1.1.0
 matplotlib==3.1.0
+monotonic==1.5
 numpy==1.16.4
 pyparsing==2.4.0
 python-dateutil==2.8.0
 python-engineio==3.8.1
 python-socketio==4.1.0
+redis==3.2.1
 six==1.12.0
 tqdm==4.32.2
