+++ frontend/app/api.ts
@@ -1,23 +1,24 @@
 import * as demofile from "demofile"
 
 export interface Header {
-    serverName: string,
-    tickRate: number
+  serverName: string,
+  tickRate: number
 }
 
 interface Socket {
-  send: (payload: any) => void
+  send: (payload: any) => void,
+  emit: (event: string, payload?: any) => void
 }
 
 interface PlayerDeathEvent {
-    userid: number,
-    attacker: number,
-    headshot: boolean,
-    weapon: string
+  userid: number,
+  attacker: number,
+  headshot: boolean,
+  weapon: string
 }
 
 interface PlayerFootstepEvent {
-    userid: number
+  userid: number
 }
 
 interface RoundEndEvent {
@@ -36,98 +37,100 @@ function stringify(event: string, payload?: any): string {
 }
 
 export class WebSocketProtocol {
-    constructor(
-        readonly socket: Socket,
-        readonly onStart: (header: Header) => void
-    ) {}
+  constructor(
+    readonly socket: Socket,
+    readonly onStart: (header: Header) => void
+  ) { }
 
-    fromBuffer(buf: ArrayBuffer) {
-        const demoFile = new demofile.DemoFile();
-        demoFile.on("start", () => {
-            this.onStart(demoFile.header)
-            this.socket.send(stringify("start", demoFile.header))
-        })
+  fromBuffer(buf: ArrayBuffer) {
+    this.socket.emit("new_upload")
+    const demoFile = new demofile.DemoFile();
+    demoFile.on("start", () => {
+      this.onStart(demoFile.header)
+      this.socket.send(stringify("start", demoFile.header))
+    })
 
-        demoFile.gameEvents.on("player_death", (e: PlayerDeathEvent) => {
-          const victim = demoFile.entities.getByUserId(e.userid)
-          const victimName = victim ? victim.name : null
-      
-          // Attacker may have disconnected so be aware.
-          // e.g. attacker could have thrown a grenade, disconnected, then that grenade
-          // killed another player.
-          const attacker = demoFile.entities.getByUserId(e.attacker)
-          const attackerName = attacker ? attacker.name : null
-      
-          this.socket.send(stringify("player_death", {
-            attacker: {
-              id: e.attacker,
-              attackerName: attackerName,
-            },
-            victim: {
-              name: victimName,
-              id: e.userid,
-            },
-            headshot: e.headshot,
-            weapon: e.weapon,
-            time: demoFile.currentTime,
-          }))
-        });
+    demoFile.gameEvents.on("player_death", (e: PlayerDeathEvent) => {
+      const victim = demoFile.entities.getByUserId(e.userid)
+      const victimName = victim ? victim.name : null
 
-        demoFile.gameEvents.on("player_footstep", (e: PlayerFootstepEvent) => {
-          const player = demoFile.entities.getByUserId(e.userid)
-          if (player) {
-            this.socket.send(stringify("player_footstep", {
-              id: e.userid,
-              position: player.position,
-              time: demoFile.currentTime,
-            }))
-          }
-        })
-      
-        demoFile.gameEvents.on("round_start", (e: RoundStartEvent)=> {
-          this.socket.send(stringify("round_start", {
-            time: demoFile.currentTime,
-            timelimit: e.timelimit,
-            fraglimit: e.fraglimit,
-            objective: e.objective
-          }))
-        })
-      
-        demoFile.gameEvents.on("round_announce_match_start", ()=> {
-          this.socket.send(stringify("round_announce_match_start", {
-            time: demoFile.currentTime
-          }))
-        })
-      
-        demoFile.gameEvents.on("round_end", (e: RoundEndEvent)=> {
-          this.socket.send(stringify("round_end", {
-            time: demoFile.currentTime,
-            reason: e.reason
-          }))
-        })
-      
-        demoFile.gameEvents.on("round_officially_ended", () => {
-          const teams = demoFile.teams
-          const ts = teams[2]
-          const cts = teams[3]
+      // Attacker may have disconnected so be aware.
+      // e.g. attacker could have thrown a grenade, disconnected, then that grenade
+      // killed another player.
+      const attacker = demoFile.entities.getByUserId(e.attacker)
+      const attackerName = attacker ? attacker.name : null
 
-          this.socket.send(stringify("round_officially_ended", {
-            ts: {
-              score: ts.score
-            },
-            cts: {
-              score: cts.score
-            }
-          }))
-        })
+      this.socket.send(stringify("player_death", {
+        attacker: {
+          id: e.attacker,
+          attackerName: attackerName,
+        },
+        victim: {
+          name: victimName,
+          id: e.userid,
+        },
+        headshot: e.headshot,
+        weapon: e.weapon,
+        time: demoFile.currentTime,
+      }))
+    });
 
-        demoFile.on("end", () => {
-          this.socket.send(stringify("end"))
-          // Stop parsing - we're finished
-          demoFile.cancel()
-        })
+    demoFile.gameEvents.on("player_footstep", (e: PlayerFootstepEvent) => {
+      const player = demoFile.entities.getByUserId(e.userid)
+      if (player) {
+        this.socket.send(stringify("player_footstep", {
+          id: e.userid,
+          position: player.position,
+          time: demoFile.currentTime,
+        }))
+      }
+    })
 
-        
-        demoFile.parse(buf);
-    }
+    demoFile.gameEvents.on("round_start", (e: RoundStartEvent) => {
+      this.socket.send(stringify("round_start", {
+        time: demoFile.currentTime,
+        timelimit: e.timelimit,
+        fraglimit: e.fraglimit,
+        objective: e.objective
+      }))
+    })
+
+    demoFile.gameEvents.on("round_announce_match_start", () => {
+      this.socket.send(stringify("round_announce_match_start", {
+        time: demoFile.currentTime
+      }))
+    })
+
+    demoFile.gameEvents.on("round_end", (e: RoundEndEvent) => {
+      this.socket.send(stringify("round_end", {
+        time: demoFile.currentTime,
+        reason: e.reason
+      }))
+    })
+
+    demoFile.gameEvents.on("round_officially_ended", () => {
+      const teams = demoFile.teams
+      const ts = teams[2]
+      const cts = teams[3]
+
+      this.socket.send(stringify("round_officially_ended", {
+        ts: {
+          score: ts.score
+        },
+        cts: {
+          score: cts.score
+        }
+      }))
+    })
+
+    demoFile.on("end", () => {
+      this.socket.send(stringify("end"))
+      this.socket.emit("render")
+      // Stop parsing - we're finished
+      demoFile.cancel()
+    })
+
+
+    demoFile.parse(buf);
+  }
 }
